moved static data to front-end app, updated .env variable
