+++ apps/front-end/.env.local
@@ -1,1 +1,1 @@
-API_URL=localhost:3333/api
+NEXT_PUBLIC_API_URL=http://localhost:3333/api
