+++ apps/front-end/pages/api/getItemsGroups.ts
@@ -1,4 +1,4 @@
-import { MOCK_ITEM_GROUPS } from '../../../../libs/mockData/items';
+import { MOCK_ITEM_GROUPS } from '../../libs/mockData/items';
 
 export default function handler(req, res) {
   res.status(200).send(MOCK_ITEM_GROUPS);
