+++ apps/front-end/pages/api/getItemsInfo.ts
@@ -1,4 +1,4 @@
-import { MOCK_ITEMS } from '../../../../libs/mockData/items';
+import { MOCK_ITEMS } from '../../libs/mockData/items';
 
 export default function handler(req, res) {
   res.status(200).send(MOCK_ITEMS);
